backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
supabase>=2.4.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
email-validator>=2.1.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
